build image without systemd
Not only does nixos not have a split systemd.lib package, it also gets
two copies of systemd in the image:
https://github.com/NixOS/nixpkgs/pull/97051
https://github.com/NixOS/nixpkgs/issues/98094

Let's not include it for now, the only detraction is that the
systemd/journald logger wont be available in uwsgi.

On the good side, the image went from 50MB to 30MB now.

The image is built against the `my-20.09` branch of nixpkgs, that is
based on the release-20.09 branch with uwsgi fixes added.