use /etc/default/tt-rss.conf for configuration, /var/lib is tmpfs
use /etc/default/tt-rss.conf instead of /etc/tt-rss/config.env for the
configuration, since the directory already exists on most distros, and
is somehow appropriate for keeping configuration of portable services.
also since tt-rss is configured via environment variables, the directory
doesn't need to be supplied to the service - a separate directory is not
needed.

also, /var/lib is now a tmpfs and an empty mount-point in the image.
thus systemd can create any directories under it. (and maybe the whole
/var needs to be like that TBD).